+++ R/build_episode.R
@@ -146,77 +146,20 @@ build_episode_md <- function(path, hash = NULL, outdir = path_built(path),
     quiet   = quiet
   )
 
-  # Build the article in a separate process via {callr}
+  # Build the article in a separate  process via {callr}
   # ==========================================================
   #
   # Note that this process can NOT use any internal functions
   sho <- !(quiet || identical(Sys.getenv("TESTTHAT"), "true"))
-  callr::r(function(path, hash, workenv, outpath, workdir, root, quiet) {
-    # Shortcut if the source is a markdown file
-    # Taken directly from tools::file_ext
-    file_ext <- function (x) {
-      pos <- regexpr("\\.([[:alnum:]]+)$", x)
-      ifelse(pos > -1L, substring(x, pos + 1L), "")
-    }
-    # Also taken directly from tools::file_path_sans_ext
-    file_path_sans_ext <- function (x) {
-      sub("([^.]+)\\.[[:alnum:]]+$", "\\1", x)
-    }
-    if (file_ext(path) == "md") {
-      file.copy(path, outpath, overwrite = TRUE)
-      return(NULL)
-    }
-    # Load required packages if it's an RMarkdown file
-    if (root != "") {
-      renv::load(root)
-      on.exit(renv::deactivate(root), add = TRUE)
-    }
-    # Set knitr options for output ---------------------------
-    ochunk <- knitr::opts_chunk$get()
-    oknit  <- knitr::opts_knit$get()
-    on.exit(knitr::opts_chunk$restore(ochunk), add = TRUE)
-    on.exit(knitr::opts_knit$restore(oknit), add = TRUE)
-
-    slug <- file_path_sans_ext(basename(outpath))
-
-    knitr::opts_chunk$set(
-      comment       = "",
-      fig.align     = "center",
-      class.output  = "output",
-      class.error   = "error",
-      class.warning = "warning",
-      class.message = "output",
-      fig.path      = file.path("fig", paste0(slug, "-rendered-"))
-    )
-
-    # Ensure HTML options like caption are respected by code chunks
-    knitr::opts_knit$set(
-      rmarkdown.pandoc.to = "markdown"
-    )
-
-    # Set the working directory -----------------------------
-    wd <- getwd()
-    on.exit(setwd(wd), add = TRUE)
-    setwd(workdir)
-
-    # Generate markdown -------------------------------------
-    res <- knitr::knit(
-      input = path,
-      output = outpath,
-      envir = workenv, 
-      quiet = quiet,
-      encoding = "UTF-8"
-    )
-
-    # write file to disk ------------------------------------
-    # writeLines(res, outpath)
-  },
-  args = args,
-  show = !quiet, 
-  spinner = sho,
-  env = c(callr::rcmd_safe_env(),
-    "RENV_PROFILE" = profile,
-    "RENV_CONFIG_CACHE_SYMLINKS" = renv_cache()))
+  callr::r(
+    func = callr_build_episode_md,
+    args = args,
+    show = !quiet,
+    spinner = sho,
+    env = c(callr::rcmd_safe_env(),
+      "RENV_PROFILE" = profile,
+      "RENV_CONFIG_CACHE_SYMLINKS" = renv_cache())
+  )
 
   invisible(outpath)
 }